% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSP_Norm_list.R
\name{DSPNorm.preListed}
\alias{DSPNorm.preListed}
\title{Normalize Data of NanoString GeoMx Protein Assays}
\usage{
DSPNorm.preListed(
  data.raw,
  area = NA,
  nuclei.count = NA,
  control.hk = c("S6", "GAPDH", "Histone H3"),
  control.neg = c("Ms IgG1", "Ms IgG2a", "Rb IgG")
)
}
\arguments{
\item{data.raw}{a raw DSP data matrix that has only been ERCC-normalized. The raw data matrix may be either a matrix or data.frame object with each row for a protein and each column for a ROI.}

\item{area}{a numerical vector of ROI areas, must be arranged in the same order of the columns of data.raw. If missing, the normalization methods concerning area cannot be performed.}

\item{nuclei.count}{a numerical vector of ROI nuclei counts, must be arranged in the same order of the columns of data.raw. If missing, the normalization methods concerning nuclei count cannot be performed.}

\item{control.hk}{a character vector listing all the housekeeping proteins that appear in the row names of data.raw.}

\item{control.neg}{a character vector listing all the negative control proteins that appear in the row names of data.raw.}
}
\value{
a list object containing the following 3 slots: 1. "norm.data.list": a list of normalized DSP data matrices by selected normalization methods; 2. "methods.available": a numerical vector of those selected method IDs; 3. "methods.table.summary": a table of data.frame class manifesting the details of all the 17 methods.
}
\description{
DSPNorm.preListed provides convenience of easily choosing among the 17 pre-defined unique normalization methods for raw data (only ERCC-normalized) of NanoString GeoMx protein assays for downstream performance evaluation. Note that those pre-defined normalization methods that are not applicable due to absence of required input info will be automatically skipped.
}
\examples{
data(COVID.19.DSP)

data.raw <- COVID.19.DSP$Expr.raw.DSP

group.info <- COVID.19.DSP$Meta.DSP

# automatically obtain normalized DSP data by the 17 built-in methods:
preListed <- DSPNorm.preListed(data.raw = data.raw,
                               area = group.info$area,
                               nuclei.count=NA,
                               control.hk=c("S6", "GAPDH", "Histone H3"),
                               control.neg=c("Ms IgG1", "Ms IgG2a", "Rb IgG"))

# the methods 6, 7, and 17 are skipped due the absence of nuclei.count:
preListed$methods.available
# [1]  1  2  3  4  5  8  9 10 11 12 13 14 15 16
}
