% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSP_Norm.R
\name{DSPNorm}
\alias{DSPNorm}
\title{Normalize Data of NanoString GeoMx Protein Assays}
\usage{
DSPNorm(
  data.raw,
  area = NA,
  nuclei.count = NA,
  norm.instruction = list(scaling = "None", bkg.removal = "None", control.neg =
    c("Ms IgG1", "Ms IgG2a", "Rb IgG"), norm.to.hk = "HK_geoMean", control.hk = c("S6",
    "GAPDH", "Histone H3"))
)
}
\arguments{
\item{data.raw}{a raw DSP data matrix that has only been ERCC-normalized. The raw data matrix may be either a matrix or data.frame object with each row for a protein and each column for a ROI.}

\item{area}{a numerical vector of ROI areas, must be arranged in the same order of the columns of data.raw. If missing, the normalization methods concerning area cannot be performed.}

\item{nuclei.count}{a numerical vector of ROI nuclei counts, must be arranged in the same order of the columns of data.raw. If missing, the normalization methods concerning nuclei count cannot be performed.}

\item{norm.instruction}{a list specifying scaling: a string of "None", "Area_geoMean" (in which case the argument area must be provided), or "Nuclei_geoMean"(in which case the argument nuclei.count must be provided); background removal: a string of "None", "STNR_geoMean" or "STNS_mean2SD" (in which case a character vector of negative controls, control.neg, must be provided); and if normalization to HK controls shall be applied: a string of "None" or "HK_geoMean" (in which case a character vector of housekeeping controls, control.hk, must be provided).}
}
\value{
a matrix or data.frame object, same class and data structure as the input of raw DSP data matrix.
}
\description{
DSPNorm provides flexible options of normalization methods for raw data (only ERCC-normalized) of NanoString GeoMx protein assays for downstream performance evaluation.
}
\examples{
data(COVID.19.DSP)

data.raw <- COVID.19.DSP$Expr.raw.DSP

group.info <- COVID.19.DSP$Meta.DSP

# normalize the raw data by area of each ROI:
data.norm <- DSPNorm(data.raw = data.raw,
                     area = group.info$area,
                     norm.instruction = list(scaling="Area_geoMean",
                                             bkg.removal="None",
                                             norm.to.hk="None"))

# normalize the raw data by geometric means of the two housekeeping controls "S6" and "GAPDH":
data.norm <- DSPNorm(data.raw = data.raw,
                     norm.instruction = list(scaling="None",
                                             bkg.removal="None",
                                             norm.to.hk="HK_geoMean",
                                             control.hk = c("S6", "GAPDH")))

# normalize the raw data first by background removal, "STNS_mean2SD", then by geometric means
# of three housekeeping controls "S6", "GAPDH", and "Histone H3":
data.norm <- DSPNorm(data.raw = data.raw,
                     norm.instruction = list(scaling="None",
                                             bkg.removal="STNS_mean2SD",
                                             control.neg=c("Ms IgG1", "Ms IgG2a", "Rb IgG"),
                                             norm.to.hk="HK_geoMean",
                                             control.hk = c("S6", "GAPDH", "Histone H3")))
}
